requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
backports.functools-lru-cache==1.6.1
bcrypt==3.2.0
brotlipy==0.7.0
bumpversion==0.5.3
cachetools==4.2.1
celery==5.2.1
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
cloudpickle==1.6.0
constantly==15.1.0
cryptography==3.4.4
cssselect==1.1.0
cssutils==1.0.2
cycler==0.10.0
cytoolz==0.11.0
daphne==3.0.1
dask==2021.2.0
decorator==4.4.2
Django==3.2
django-braces==1.13.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.4.0
django-environ==0.4.5
django-guardian==2.1.0
django-oauth-toolkit==1.4.0
django-timezone-field==4.1.2
django-utils-six==2.0
djangorestframework==3.12.2
environs==9.3.1
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imagecodecs==2021.1.11
imageio==2.9.0
incremental==17.5.0
kiwisolver==1.3.1
lxml==4.6.2
marshmallow==3.10.0
matplotlib==3.3.4
mock==4.0.3
msgpack==1.0.2
networkx==2.5
numpy==1.20.1
oauthlib==3.1.0
olefile==0.46
packaging==20.9
pandas==1.2.2
Pillow==8.1.0
pip
pooch==1.3.0
premailer==3.7.0
priority==1.3.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
qrcode==6.1
redis==3.5.3
reportlab==3.5.66
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.0
sentry-sdk==0.20.3
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
tifffile==2021.2.1
toolz==0.11.1
tornado==6.1
Twisted
txaio==21.2.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
wheel==0.36.2
whitenoise==5.2.0
xlrd==2.0.1
yagmail==0.14.245
zope.interface==5.2.0
torch
torchvision
opencv-contrib-python
gunicorn
seaborn
gviz_api
flower
htmlmin
drf-jwt
pyjwt
